chore: migrate tests to pytest, add ruff + mypy, fill test gaps
- 迁移 8 个手写测试脚本为 pytest（扫描/分组/国际化/移动/磁盘/冒烟/UI状态）
- 改 perf_test.py → perf_probe.py（采样脚本非测试）
- 新增 5 个测试文件：filters / config / storage / panel / settings
- 配置 ruff（100列/3.10）排除已有惯例（BLE001/DTZ/S110/B008/SIM102等）
- 配置 mypy 排除 PySide6 类型桩误报；定向忽略 widget/panel 动态类型
- ball.py 移除多余 int(round(...)) 调用
- 61 个测试全部通过，ruff 零错误，mypy 17 文件零问题

+++ diskwatch/ui/ball.py
@@ -94,7 +94,7 @@ class MiniBall(QWidget):
         self.update()
 
     def _update_tooltip(self) -> None:
-        pct = int(round(self._ratio * 100))
+        pct = round(self._ratio * 100)
         self.setToolTip(
             tr(
                 "今日 {today}\n近{days}天合计 {total}\n今日占比 {pct}%",
